Refactor TaskManager
